Keep actual order, move tests to unit
